crypto_trading_system/requirements.txt
python-binance==1.0.17
aiohttp>=3.9,<4
websockets>=12,<13
